backend/pyproject.toml
[project]
name = "don-padel-backend"
version = "0.1.0"
description = "Backend API and ingestion for Swiss Playtomic padel availability."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.16.0",
  "apscheduler>=3.10.4",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "pydantic-settings>=2.9.0",
  "python-dateutil>=2.9.0.post0",
  "sqlalchemy>=2.0.0",
  "uvicorn[standard]>=0.34.0",
  "psycopg[binary]>=3.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
